requirements.txt
flask~=2.3.3
requests~=2.31.0
bcrypt~=4.0.1
Faker~=19.6.2
python-dotenv~=1.0.0
mysql-connector-python~=8.1.0